requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
PyJWT==1.7.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
